+++ connectors/hazelcast/hazelcast-connector/src/main/java/org/platformlambda/hazelcast/reporter/PresenceSetup.java
@@ -24,6 +24,7 @@ import org.platformlambda.core.system.PostOffice;
 import org.platformlambda.core.util.AppConfigReader;
 import org.platformlambda.core.util.Utility;
 import org.platformlambda.core.websocket.client.PersistentWsClient;
+import org.platformlambda.hazelcast.TopicLifecycleListener;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -51,7 +52,9 @@ public class PresenceSetup implements CloudSetup {
                 try {
                     String url = reader.getProperty(PRESENCE_MONITOR, "ws://127.0.0.1:8080/ws/presence");
                     List<String> monitors = Utility.getInstance().split(url, ", ");
-                    new PersistentWsClient(PresenceConnector.getInstance(), monitors).start();
+                    PersistentWsClient ws = new PersistentWsClient(PresenceConnector.getInstance(), monitors);
+                    ws.setCondition(TopicLifecycleListener::isReady);
+                    ws.start();
                 } catch (Exception e) {
                     log.error("Unable to start", e);
                     System.exit(-1);